unsupervised_strategy/requirements.txt
absl-py==0.9.0
astor==0.8.1
cycler==0.10.0
gast==0.3.3
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.3
mtcnn==0.1.0
numpy==1.15.0
opencv-python==4.2.0.32
pandas==1.0.1
Pillow==7.0.0
pkg-resources==0.0.0
protobuf==3.11.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tqdm==4.42.1
Werkzeug==0.16.1
wrapt==1.11.2
